requirements_media_ocr_upgraded.txt
# WhatsApp Media OCR Processing Libraries (Upgraded)
# Core OCR Engines
easyocr>=1.7.0
PyMuPDF>=1.23.0
pdf2image>=1.16.3
pytesseract>=0.3.10
Pillow>=10.0.0



# Google Cloud Vision & Document AI
google-cloud-vision>=3.4.0
google-cloud-documentai>=2.20.0

# HTTP Requests (for API calls)
requests>=2.31.0

# Azure Vision API (Legacy)
azure-cognitiveservices-vision-computervision>=0.9.0
msrest>=0.7.1

# Existing MACHO-GPT Dependencies
playwright>=1.40.0
openai>=1.0.0
pyyaml>=6.0
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.17.0

# Additional Utilities
click>=8.1.0
python-dotenv>=1.0.0
numpy<2.0.0
opencv-python-headless>=4.8.0 